requirements.txt
streamlit==1.25.0
pandas==1.5.3
ultralytics==8.0.134
openai==0.27.8
python-dotenv==1.0.0
